extract history view from storm
